fix: include Undetermined Indices in AVITI summary emails
undReads was computed via round(x, 0), which returns a float in
Python. The email-body logic in flowcell.py only special-cases str
(multi-lane) and int (single-lane) undetermined values, so the float
case silently fell through neither branch and AVITI runs never got
an "Undetermined indices" line.

+++ src/dissectBCL/fakeNews.py
@@ -402,7 +402,7 @@ def gatherFinalMetrics(outLane, flowcell):
         with open(outPath / "RunStats.json") as f:
             _rundata = json.load(f)
         totalReads = _rundata["NumPolonies"]
-        undReads = round((_rundata["PercentAssignedReads"] / 100) * totalReads, 0)
+        undReads = int(round((_rundata["PercentAssignedReads"] / 100) * totalReads, 0))
         # topBarcodes
         bcDF = pd.read_csv(outPath / "UnassignedSequences.csv")
         tot_und = bcDF["Count"].sum()
